Use Google's CDN for how-jquery-works examples
It's both best practice and easier for a user to try out as they don't
need to navigate downloading jQuery and saving it to the right place.

I've included the http: rather than start with // because people are
likely to run the tests on file:// URLs, and that case should work.
People are unlikely to try this out on https:// URLs, so there should
be no problem forcing http.