Improve error callback example for $.ajax
The error handler includes two of the three parameters, but doesn't actually do anything with them. This is a very common place where new developers will get the "Sorry, there was a problem!" and not know how to go further to figure out what the problem was. Adding the logging may help improve that.