Removed reference to use of hungarian notation.
Originally the variable was called $divs (I think). However the use of Hungarian notation has since been removed from this tutorial.

I noticed while reading this that there was still an explanation as to why the variable started with a "$" although the variable no longer did. This kind of confusing so I've removed it and I think it's better without the explanation.