Made text reflect code. Set "rel" to valid value.
In the code "html:" is used. However in the text underneath "text" is mentioned.
This is now corrected to talk about "html" underneath instead.

Also. Previously the tutorial sets a link's "rel" to "super-special". As
"super-special" isn't a valid value for "rel" I have changed it to "nofollow"
as this is a valid value (and fairly well known). This is really just a taste
thing but I found "super-special" a little confusing here.