Replacing references to deprecated methods
I'm learning jQ,  so this is by no means going to be  comprehensive but as of the docs:

```
Deprecation Notice: The jqXHR.success(), jqXHR.error(), and jqXHR.complete() callbacks are deprecated as of jQuery 1.8. To prepare your code for their eventual removal, use jqXHR.done(), jqXHR.fail(), and jqXHR.always() instead.
```

I have made changes to reflect these.