Incorrect explanation. Triggering event isn't an anchor Tag.
The triggering element has a tag which is an anchor tag, in other words, the triggering element's tag is an anchor tag.

Some possible information to add?
The event is Clicking on the element. It's the first element in the bubbling "chain". One could call event.stopPropagation to stop the event from bubbling up the DOM tree.